rag-4/pinecone/pyproject.toml
[tool.poetry]
name = "rag-chain"
version = "0.1.1"
description = ""
authors = ["Lance Martin <lance@langchain.dev>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
pinecone-client = "3.0.0"
openai = "^1.6.1"
python-dotenv = "^1.0.0"
# langchain-community = ">=0.0.13,<0.1"
jupyter = "^1.0.0"
langchain-community = "*"
langchain-openai = "*"
langchain-core = "*"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
